Нашла решениеДЗ

+++ Урок1/Domzad/Program.cs
@@ -2,6 +2,6 @@
 long b = 2;
 long c = 4;
 int d = 3;
-double Formu = a*b/(c+d);
+double Formu = (double)a*b/(c+d);
 Console.WriteLine(Formu);
